+++ core/src/main/java/org/fao/geonet/kernel/datamanager/draft/DraftMetadataManager.java
@@ -2,8 +2,10 @@ package org.fao.geonet.kernel.datamanager.draft;
 
 import javax.annotation.Nonnull;
 import javax.annotation.PostConstruct;
+import javax.persistence.EntityNotFoundException;
 
 import org.apache.commons.lang.NotImplementedException;
+import org.fao.geonet.constants.Geonet;
 import org.fao.geonet.domain.IMetadata;
 import org.fao.geonet.domain.Metadata;
 import org.fao.geonet.domain.MetadataDraft;
@@ -12,8 +14,13 @@ import org.fao.geonet.kernel.datamanager.base.BaseMetadataManager;
 import org.fao.geonet.repository.MetadataDraftRepository;
 import org.fao.geonet.repository.PathSpec;
 import org.fao.geonet.repository.Updater;
+import org.fao.geonet.utils.Log;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import jeeves.server.context.ServiceContext;
 
@@ -41,26 +48,34 @@ public class DraftMetadataManager extends BaseMetadataManager implements IMetada
      */
     @Override
     public synchronized void updateMetadataOwner(final int id, final String owner, final String groupOwner) throws Exception {
-        metadataDraftRepository.update(id, new Updater<MetadataDraft>() {
-            @Override
-            public void apply(@Nonnull MetadataDraft entity) {
-                entity.getSourceInfo().setGroupOwner(Integer.valueOf(groupOwner));
-                entity.getSourceInfo().setOwner(Integer.valueOf(owner));
-            }
-        });
-
-        super.updateMetadataOwner(id, owner, groupOwner);
+        try {
+            metadataDraftRepository.update(id, new Updater<MetadataDraft>() {
+                @Override
+                public void apply(@Nonnull MetadataDraft entity) {
+                    entity.getSourceInfo().setGroupOwner(Integer.valueOf(groupOwner));
+                    entity.getSourceInfo().setOwner(Integer.valueOf(owner));
+                }
+            });
+        } catch (EntityNotFoundException e) {
+            super.updateMetadataOwner(id, owner, groupOwner);
+        }
     }
 
     @Override
+    @Transactional(readOnly=false, isolation=Isolation.READ_COMMITTED, propagation=Propagation.REQUIRES_NEW)
     public IMetadata save(IMetadata info) {
         if (info instanceof Metadata) {
             return super.save((Metadata) info);
         } else if (info instanceof MetadataDraft) {
-            return metadataDraftRepository.save((MetadataDraft) info);
+            try {
+                return metadataDraftRepository.save((MetadataDraft) info);
+            } catch (Throwable t) {
+                Log.error(Geonet.DATA_MANAGER, t.getMessage(), t);
+            }
         } else {
             throw new NotImplementedException("Unknown IMetadata subtype: " + info.getClass().getName());
         }
+        return info;
     }
 
     @SuppressWarnings("unchecked")
@@ -82,9 +97,12 @@ public class DraftMetadataManager extends BaseMetadataManager implements IMetada
     }
 
     @Override
+    @Transactional(readOnly = false, isolation = Isolation.READ_COMMITTED, noRollbackFor = { EmptyResultDataAccessException.class })
     public void delete(Integer id) {
-        metadataDraftRepository.delete(id);
         super.delete(id);
+        if (metadataDraftRepository.exists(id)) {
+            metadataDraftRepository.delete(id);
+        }
     }
 
     @SuppressWarnings("unchecked")
